judge/pyproject.toml
[project]
name = "judge-agent"
version = "0.1.0"
description = "Judge microservice agent"
requires-python = ">=3.10,<3.14"
dependencies = [
    "google-adk>=0.1.4",
    "a2a-sdk==0.3.22",
    "uvicorn",
    "fastapi==0.128.0",
    "google-cloud-logging==3.12.0",
    "opentelemetry-exporter-gcp-trace==1.9.0",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
